Minor misc revisions to `URITemplate`
